+++ src/themachinethatgoesping/echosounders/em3000/datagrams/surfacesoundspeeddatagram.hpp
@@ -68,7 +68,7 @@ class SurfaceSoundSpeedDatagram : public EM3000Datagram
     // getters
     uint16_t get_sound_speed_counter() const { return _sound_speed_counter; }
     uint16_t get_system_serial_number() const { return _system_serial_number; }
-    int32_t  get_number_of_entries() const { return _number_of_entries; }
+    uint16_t  get_number_of_entries() const { return _number_of_entries; }
     uint8_t  get_spare() const { return _spare; }
     uint8_t  get_etx() const { return _etx; }
     uint16_t get_checksum() const { return _checksum; }
